[Array2D] some fixes
